Extended holes/placeholders to args and kwargs
Previously it was only applicable to args, now it happens with kwargs
too

One peculiarity with kwargs(unlike args) is that they can be
overwritten.

For example if a function has 2 kwargs named `foo` and `bar`, you could
potentially hole `foo` and supply `bar` but on next call, supply `foo`
but overwrite `bar`. This results in the exected still holed-function.