Introducing tags to zephyr
Tags will enable users to do contextualized updates to leaves of params
(or any pytree dicts). They can use tags to decide how to update the
leaves.

Tags are just the keys in the dict of params. <- this means that naming
the keys can be much more meaningful.

In the future (coming to zephyr):
you could tag certain things as "A" and others "B" and update those
"A"'s differentely from those tagged with "B"'s